Added mixin for parameters. Added estimator cloning based on params. Simplified proximity class hierarchy
