+++ src/tdamapper/cover.py
@@ -6,78 +6,42 @@ the whole dataset. Unlike clustering, open subsets do not need to be disjoint.
 Indeed, the overlaps of the open subsets define the edges of the Mapper graph.
 """
 
-from tdamapper.proximity import (
-    proximity_net,
-    BallProximity,
-    KNNProximity,
-    CubicalProximity,
-    TrivialProximity
-)
-from tdamapper._common import ParamsMixin
+import numpy as np
+
+from tdamapper.core import Proximity
+from tdamapper.utils.metrics import get_metric, chebyshev
+from tdamapper.utils.vptree import VPTree
 
 
-class Cover(ParamsMixin):
+class _Pullback:
+
+    def __init__(self, fun, dist):
+        self.fun = fun
+        self.dist = dist
+
+    def __call__(self, x, y):
+        return self.dist(self.fun(x), self.fun(y))
+
+
+def _snd(x):
+    return x[1]
+
+
+def _rho(x):
+    return np.floor(x) + 0.5
+
+
+class BallCover(Proximity):
     """
-    Abstract interface for cover algorithms.
+    Cover algorithm based on `ball proximity function`, covering data with
+    open balls of fixed radius.
 
-    This is a naive implementation. Subclasses should override the methods of
-    this class to implement more meaningful cover algorithms.
-    """
-
-    def apply(self, X):
-        """
-        Covers the dataset with a single open set.
-
-        This is a naive implementation that should be overridden by subclasses
-        to implement more meaningful cover algorithms.
-
-        :param X: A dataset of n points to be covered with open subsets.
-        :type X: array-like of shape (n, m) or list-like of length n
-        :return: A generator that produces a single list of ints whose elements
-            are the indices of the data points, ranging from 0 to n - 1.
-        :rtype: generator of lists of ints
-        """
-        yield list(range(0, len(X)))
-
-
-class ProximityCover(Cover):
-    """
-    Cover algorithm based on proximity-net.
-
-    This class creates an open cover by calling the function
-    :func:`tdamapper.proximity.proximity_net`.
-
-    :param proximity: The proximity function to use.
-    :type proximity: Anything compatible with class
-        :class:`tdamapper.proximity.Proximity`
-    """
-
-    def __init__(self, proximity):
-        self.__proximity = proximity
-
-    def apply(self, X):
-        """
-        Covers the dataset using proximity-net on the specified proximity
-        function.
-
-        The proximity function is used to create an open set whenever a point
-        is picked from :func:`tdamapper.proximity.proximity_net`.
-
-        :param X: A dataset of n points to be covered with open subsets.
-        :type X: array-like of shape (n, m) or list-like of length n
-        :return: A generator yielding lists of ints whose elements are the
-            indices of the data points.
-        :rtype: generator of lists of ints
-        """
-        return proximity_net(X, self.__proximity)
-
-
-class BallCover(ProximityCover):
-    """
-    Cover algorithm based on `ball proximity function` implemented as
-    :class:`tdamapper.proximity.BallProximity`.
+    An open ball is a set of points within a specified distance from a center
+    point. This class maps each point to its corresponding open ball with a
+    fixed radius centered on the point itself.
 
     :param radius: The radius of the open balls. Must be a positive value.
+        Defaults to 1.0.
     :type radius: float
     :param metric: The metric used to define the distance between points.
         Accepts any value compatible with `tdamapper.utils.metrics.get_metric`.
@@ -104,7 +68,7 @@ class BallCover(ProximityCover):
 
     def __init__(
         self,
-        radius,
+        radius=1.0,
         metric='euclidean',
         metric_params=None,
         kind='flat',
@@ -112,25 +76,74 @@ class BallCover(ProximityCover):
         leaf_radius=None,
         pivoting=None,
     ):
-        prox = BallProximity(
-            radius=radius,
-            metric=metric,
-            metric_params=metric_params,
-            kind=kind,
-            leaf_capacity=leaf_capacity,
-            leaf_radius=leaf_radius,
-            pivoting=pivoting,
+        self.radius = radius
+        self.metric = metric
+        self.metric_params = metric_params
+        self.kind = kind
+        self.leaf_capacity = leaf_capacity
+        self.leaf_radius = leaf_radius
+        self.pivoting = pivoting
+
+    def fit(self, X):
+        """
+        Train internal parameters.
+
+        This method creates a vptree on the dataset in order to perform fast
+        range queries in the func:`tdamapper.cover.BallCover.search`
+        method.
+
+        :param X: A dataset of n points used to extract parameters and perform
+            training.
+        :type X: array-like of shape (n, m) or list-like of length n
+        :return: The object itself.
+        :rtype: self
+        """
+        metric = get_metric(self.metric, **(self.metric_params or {}))
+        self.__radius = self.radius
+        self.__data = list(enumerate(X))
+        self.__vptree = VPTree(
+            self.__data,
+            metric=_Pullback(_snd, metric),
+            metric_params=None,
+            kind=self.kind,
+            leaf_capacity=self.leaf_capacity,
+            leaf_radius=self.leaf_radius or self.radius,
+            pivoting=self.pivoting,
         )
-        super().__init__(proximity=prox)
+        return self
+
+    def search(self, x):
+        """
+        Return a list of neighbors for the query point.
+
+        This method uses the internal vptree to perform fast range queries.
+
+        :param x: A query point for which we want to find neighbors.
+        :type x: Any
+        :return: The indices of the neighbors contained in the dataset.
+        :rtype: list[int]
+        """
+        if self.__vptree is None:
+            return []
+        neighs = self.__vptree.ball_search(
+            (-1, x),
+            self.__radius,
+            inclusive=False
+        )
+        return [x for (x, _) in neighs]
 
 
-class KNNCover(ProximityCover):
+class KNNCover(Proximity):
     """
-    Cover algorithm based on `knn proximity function` implemented as
-    :class:`tdamapper.proximity.KNNProximity`.
+    Cover algorithm based on `KNN proximity function`, covering data using
+    k-nearest neighbors (KNN).
+
+    This class maps each point to the set of the k nearest neighbors to the
+    point itself.
 
     :param neighbors: The number of neighbors to use for the KNN Proximity
         function, must be positive and less than the length of the dataset.
+        Defaults to 1.
     :type neighbors: int
     :param metric: The metric used to define the distance between points.
         Accepts any value compatible with `tdamapper.utils.metrics.get_metric`.
@@ -157,7 +170,7 @@ class KNNCover(ProximityCover):
 
     def __init__(
         self,
-        neighbors,
+        neighbors=1,
         metric='euclidean',
         metric_params=None,
         kind='flat',
@@ -165,28 +178,76 @@ class KNNCover(ProximityCover):
         leaf_radius=0.0,
         pivoting=None,
     ):
-        prox = KNNProximity(
-            neighbors=neighbors,
-            metric=metric,
-            metric_params=metric_params,
-            kind=kind,
-            leaf_capacity=leaf_capacity,
-            leaf_radius=leaf_radius,
-            pivoting=pivoting,
+        self.neighbors = neighbors
+        self.metric = metric
+        self.metric_params = metric_params
+        self.kind = kind
+        self.leaf_capacity = leaf_capacity
+        self.leaf_radius = leaf_radius
+        self.pivoting = pivoting
+
+    def fit(self, X):
+        """
+        Train internal parameters.
+
+        This method creates a vptree on the dataset in order to perform fast
+        KNN queries in the func:`tdamapper.cover.BallCover.search`
+        method.
+
+        :param X: A dataset of n points used to extract parameters and perform
+            training.
+        :type X: array-like of shape (n, m) or list-like of length n
+        :return: The object itself.
+        :rtype: self
+        """
+        metric = get_metric(self.metric, **(self.metric_params or {}))
+        self.__neighbors = self.neighbors
+        self.__data = list(enumerate(X))
+        self.__vptree = VPTree(
+            self.__data,
+            metric=_Pullback(_snd, metric),
+            metric_params=None,
+            kind=self.kind,
+            leaf_capacity=self.leaf_capacity or self.neighbors,
+            leaf_radius=self.leaf_radius,
+            pivoting=self.pivoting,
         )
-        super().__init__(proximity=prox)
+        return self
+
+    def search(self, x):
+        """
+        Return a list of neighbors for the query point.
+
+        This method queries the internal vptree in order to perform fast KNN
+        queries.
+
+        :param x: A query point for which we want to find neighbors.
+        :type x: Any
+        :return: The indices of the neighbors contained in the dataset.
+        :rtype: list[int]
+        """
+        if self.__vptree is None:
+            return []
+        neighs = self.__vptree.knn_search((-1, x), self.__neighbors)
+        return [x for (x, _) in neighs]
 
 
-class CubicalCover(ProximityCover):
+class CubicalCover(Proximity):
     """
-    Cover algorithm based on the `cubical proximity function` implemented as
-    :class:`tdamapper.proximity.CubicalProximity`.
+    Cover algorithm based on the `cubical proximity function`, covering data
+    with open hypercubes of uniform size and overlap.
+
+    A hypercube is a multidimensional generalization of a square or a cube.
+    The size and overlap of the hypercubes are determined by the number of
+    intervals and the overlap fraction parameters. This class maps each point
+    to the hypercube with the nearest center.
 
     :param n_intervals: The number of intervals to use for each dimension.
         Must be positive and less than or equal to the length of the dataset.
+        Defaults to 1.
     :type n_intervals: int
     :param overlap_frac: The fraction of overlap between adjacent intervals on
-        each dimension, must be in the range (0.0, 1.0).
+        each dimension, must be in the range (0.0, 1.0). Defaults to 0.5.
     :type overlap_frac: float
     :param metric: The metric used to define the distance between points.
         Accepts any value compatible with `tdamapper.utils.metrics.get_metric`.
@@ -196,8 +257,8 @@ class CubicalCover(ProximityCover):
         passed to `tdamapper.utils.metrics.get_metric`. Defaults to None.
     :type metric_params: dict, optional
     :param kind: Specifies whether to use a flat or a hierarchical vantage
-        point tree. Acceptable values are 'flat' or 'hierarchical'. Defaults
-        to 'flat'.
+        point tree. Acceptable values are 'flat' or 'hierarchical'. Defaults to
+        'flat'.
     :type kind: str
     :param leaf_capacity: The maximum number of points in a leaf node of the
         vantage point tree. Must be a positive value. Defaults to 1.
@@ -213,29 +274,84 @@ class CubicalCover(ProximityCover):
 
     def __init__(
         self,
-        n_intervals,
-        overlap_frac,
+        n_intervals=1,
+        overlap_frac=0.5,
         kind='flat',
         leaf_capacity=1,
         leaf_radius=None,
         pivoting=None,
     ):
-        prox = CubicalProximity(
-            n_intervals=n_intervals,
-            overlap_frac=overlap_frac,
-            kind=kind,
-            leaf_capacity=leaf_capacity,
-            leaf_radius=leaf_radius,
-            pivoting=pivoting,
+        self.n_intervals = n_intervals
+        self.overlap_frac = overlap_frac
+        self.kind = kind
+        self.leaf_capacity = leaf_capacity
+        self.leaf_radius = leaf_radius
+        self.pivoting = pivoting
+
+    def _gamma_n(self, x):
+        return self.__n_intervals * (x - self.__min) / self.__delta
+
+    def _gamma_n_inv(self, x):
+        return self.__min + self.__delta * x / self.__n_intervals
+
+    def _phi(self, x):
+        return self._gamma_n_inv(_rho(self._gamma_n(x)))
+
+    def _get_bounds(self, data):
+        if (data is None) or len(data) == 0:
+            return
+        minimum, maximum = data[0], data[0]
+        eps = np.finfo(np.float64).eps
+        for w in data:
+            minimum = np.minimum(minimum, np.array(w))
+            maximum = np.maximum(maximum, np.array(w))
+        _min = np.nan_to_num(minimum, nan=-float(eps))
+        _max = np.nan_to_num(maximum, nan=float(eps))
+        _delta = np.maximum(eps, _max - _min)
+        return _min, _max, _delta
+
+    def _convert(self, X):
+        return np.asarray(X).reshape(len(X), -1).astype(float)
+
+    def fit(self, X):
+        """
+        Train internal parameters.
+
+        This method builds an internal :class:`tdamapper.cover.BallCover`
+        attribute that allows efficient queries of the dataset.
+
+        :param X: A dataset of n points used to extract parameters and perform
+            training.
+        :type X: array-like of shape (n, m) or list-like of length n
+        :return: The object itself.
+        :rtype: self
+        """
+        self.__overlap_frac = self.overlap_frac
+        self.__n_intervals = self.n_intervals
+        self.__radius = 1.0 / (2.0 - 2.0 * self.__overlap_frac)
+        XX = self._convert(X)
+        self.__ball_proximity = BallCover(
+            self.__radius,
+            metric=_Pullback(self._gamma_n, chebyshev()),
+            kind=self.kind,
+            leaf_capacity=self.leaf_capacity,
+            leaf_radius=self.leaf_radius,
+            pivoting=self.pivoting
         )
-        super().__init__(proximity=prox)
+        self.__min, self.__max, self.__delta = self._get_bounds(XX)
+        self.__ball_proximity.fit(XX)
+        return self
 
+    def search(self, x):
+        """
+        Return a list of neighbors for the query point.
 
-class TrivialCover(ProximityCover):
-    """
-    Cover algorithm based on the `trivial proximity function` implemented as
-    :class:`tdamapper.proximity.TrivialProximity`.
-    """
+        This method takes a target point as input and returns the hypercube
+        whose center is closest to the target point.
 
-    def __init__(self):
-        super().__init__(proximity=TrivialProximity())
+        :param x: A query point for which we want to find neighbors.
+        :type x: Any
+        :return: The indices of the neighbors contained in the dataset.
+        :rtype: list[int]
+        """
+        return self.__ball_proximity.search(self._phi(x))
