+++ tests/test_unit_knn.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 
-from tdamapper.proximity import KNNProximity
+from tdamapper.cover import KNNCover
 from tdamapper.utils.metrics import euclidean
 from tdamapper.utils.vptree_flat import VPTree
 
@@ -96,9 +96,9 @@ x = np.array([ 99.73199663, 100.8024564 ])
 class TestKNN(unittest.TestCase):
 
     def test_knn_search(self):
-        knn_prox = KNNProximity(neighbors=5, metric='euclidean')
-        knn_prox.fit(X)
-        neigh_ids = knn_prox.search(x)
+        knn_cover = KNNCover(neighbors=5, metric='euclidean')
+        knn_cover.fit(X)
+        neigh_ids = knn_cover.search(x)
         d = euclidean()
         dists = [d(x, X[j]) for j in neigh_ids]
         x_dist = d(x, X[5])
